+++ src/main/java/bank/recommendationservice/fintech/repository/RecommendationsRepository_PreTest.java
@@ -1,0 +1,38 @@
+package bank.recommendationservice.fintech.repository;
+
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public class RecommendationsRepository_PreTest {
+    /**
+     * Это тестовый репозиторий, предназначенный для проверки работы JDBC,
+     * в нем содержится пример sql запроса для BD
+     * @author Bogomolova
+     */
+    private final JdbcTemplate jdbcTemplate;
+
+    public RecommendationsRepository_PreTest(@Qualifier("recommendationsJdbcTemplate") JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public int getRandomTransactionAmount(UUID user){
+        Integer result = jdbcTemplate.queryForObject(
+                "SELECT amount FROM transactions t WHERE t.user_id = ? LIMIT 1",
+                Integer.class,
+                user);
+        return result != null ? result : 0;
+    }
+    public boolean checkConnection() {
+        try {
+            jdbcTemplate.queryForObject("SELECT 1", Integer.class);
+            return true;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+}
