+++ src/main/resources/application.properties
@@ -1,6 +1,5 @@
 spring.application.name=fintech
-application.fintech_service-db.url=jdbc:h2:file:./transaction.mv.db;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;TRACE_LEVEL_FILE=0
+application.fintech_service-db.url=jdbc:h2:file:./src/main/resources/transaction
 spring.datasource.driver-class-name=org.h2.Driver
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
-
