+++ src/test/java/bank/recommendationservice/fintech/repository/RecommendationsRepositoryTest.java
@@ -20,7 +20,7 @@ class RecommendationsRepositoryTest {
     private JdbcTemplate jdbcTemplate;
 
     @InjectMocks
-    private RecommendationsRepository recommendationsRepository;
+    private RecommendationsRepository_PreTest recommendationsRepository;
 
     @BeforeEach
     public void setUp() {
